# strict/swagger.json
{
    "openapi": "3.0.3",
    "paths": {
...
        "/loyer/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "LoyerController"
                ]
            }
        },
...
        "/loyer/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "LoyerController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "LoyerController"
                ]
            }
        },
        "/loyer": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "LoyerController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "datePayement": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "montant": {
                                        "type": "number"
                                    },
                                    "mois": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "LoyerController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "datePayement": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "montant": {
                                        "type": "number"
                                    },
                                    "mois": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/marqueVoiture/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MarqueVoitureController"
                ]
            }
        },
        "/marqueVoiture/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MarqueVoitureController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MarqueVoitureController"
                ]
            }
        },
        "/marqueVoiture": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MarqueVoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marqueName": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MarqueVoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "marqueName": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/voiture/option/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"[{\\\"$match\\\":{}}]\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            },
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "properties": {
                                "_bsontype": {
                                    "type": "object",
                                    "example": "\"ObjectID\""
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ]
            }
        },
        "/voiture": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "VoitureController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "year": {
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "string"
                                    },
                                    "clientId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    },
                                    "modelVoitureId": {
                                        "properties": {
                                            "_bsontype": {
                                                "type": "object",
                                                "example": "\"ObjectID\""
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}